7:4 hamming codes
This version, specifically, has the encoder and decoder interacting (one module provides the input while the other spits the output). The visible output in the FPGA is LEDs that reflect the original message encoded by the sender (after it has been decoded and corrected).

I also have isolated encoder and decoder modules for the purpose of observing their individual function/behavior on the FPGA LEDs.